pyproject.toml
[project]
name = "cunhaobot"
version = "0.1.0"
description = "A multi-platform chatbot for cuñao phrases."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=20.0",
    "google-cloud-datastore",
    "flask",
    "boto3",
    "google-cloud-storage",
    "pytz",
    "fuzzywuzzy",
    "pillow>=11.0.0",
    "tweepy",
    "gunicorn",
    "urllib3<2",
    "asgiref",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest",
    "pytest-asyncio",
    "ruff>=0.14.10",
    "ty>=0.0.7",
]

[tool.ty]
[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
exclude = ["**/tests/**"]

[tool.ty.rules]
possibly-missing-attribute = "warn"
invalid-assignment = "warn"
unresolved-attribute = "warn"
invalid-return-type = "warn"
invalid-type-form = "warn"
missing-argument = "warn"
unsupported-operator = "warn"
invalid-raise = "warn"
unknown-argument = "error"
invalid-argument-type = "warn"
